Instead of MobileLink and NavLink, now NavLink handles with responsive design + Componentizing Navbar + Removing unsued const + Removing code duplication
