+++ src/components/navigation/Navbar.js
@@ -1,11 +1,7 @@
-import React, { useCallback, useContext, useState } from "react"
+import React from "react"
 import styled from "styled-components"
-import { UserDropdown } from "./desktop/UserDropdown"
-import { Widget } from "near-social-vm"
-import { useBosLoaderStore } from "@/stores/bos-loader"
-import { AppContext } from "@/context/AppContext"
-import { navbarNavigationItems } from "@/utils/constants"
-import { mobileNavbarNavigationItems } from "utils/constants"
+import { Navigation } from "@/components/Navbar/Navigation"
+import { SigninOrJoinNow } from "components/Navbar/SinginOrJoinNow"
 
 const StyledNavbar = styled.nav`
   display: flex;
@@ -15,9 +11,6 @@ const StyledNavbar = styled.nav`
   justify-content: space-between;
 
   .logo {
-    flex-grow: 1;
-    flex-basis: 0;
-
     img {
       width: auto;
       height: 32px;
@@ -69,86 +62,11 @@ const StyledNavbar = styled.nav`
   }
 `;
 
-const NavLink = styled.a`
-  all: unset;
-
-  display: flex;
-  padding: 10px 20px;
-  justify-content: center;
-  align-items: center;
-  gap: 10px;
-
-  border-radius: 8px;
-  background: var(--bg-2, #23242b) !important;
-  color: var(--white-100, #fff) !important;
-
-  /* Other/Button_text */
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-
-  cursor: pointer;
-  transition: all 300ms;
-
-  &:hover {
-    text-decoration: none;
-    background: #ffaf51 !important;
-    color: #000 !important;
-  }
-`;
-
-const MobileDropdownButton = styled.button`
-  all: unset;
-
-  color: #fff;
-  padding: 0.5rem;
-  font-size: 2rem;
-`;
-
-const MobileLink = styled.a`
-  all: unset;
-
-  display: flex;
-  padding: 10px 20px;
-  justify-content: center;
-  align-items: center;
-  gap: 4px;
-
-  border-radius: 8px;
-  border: 1px solid var(--white-100, #fff);
-
-  color: var(--white-100, #fff);
-  transition: all 300ms;
-
-  &:hover {
-    text-decoration: none;
-    background: #fff;
-    color: #000;
-  }
-
-  cursor: pointer;
-
-  /* Other/Button_text */
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-`;
 
 const logoLink =
   "https://ipfs.near.social/ipfs/bafkreihbwho3qfvnu4yss3eh5jrx6uxhrlzdgtdjyzyjrpa6odro6wdxya";
 
 export function Navbar() {
-  const redirectStore = useBosLoaderStore()
-  const [dropdown, setDropdown] = useState(false)
-
-  const props = useContext(AppContext)
-
-  const toggleDropdown = useCallback(() => {
-    setDropdown((prev) => !prev);
-  }, [])
-
   return (
     <div
       style={{
@@ -162,92 +80,9 @@ export function Navbar() {
             <img src={logoLink} />
           </a>
         </div>
-        <div className="d-none d-md-flex flex-grow-1 flex-shrink-1 justify-content-center align-items-center gap-3 list-unstyled">
-          {navbarNavigationItems.map(({ href, label }) => {
-            const isActive = window.location.href === window.location.origin.concat(href)
-
-            return (
-              <NavLink
-                key={label}
-                href={href}
-                className={isActive && "active"}
-              >
-                {label}
-              </NavLink>
-            )
-          })}
-        </div>
-        <div className="d-none d-md-block flex-grow-1" style={{ flexBasis: 0 }}>
-          {!props.signedIn && (
-            <button
-              className="sign-in"
-              style={{ width: "max-content", marginLeft: "auto" }}
-              onClick={props.requestSignIn}
-            >
-              Sign In
-            </button>
-          )}
-          {props.signedIn && (
-            <div>
-              <Widget
-                src="buildhub.near/widget/components.button.join-now"
-                config={{
-                  redirectMap: redirectStore.redirectMap,
-                }}
-                props={{
-                  children: <UserDropdown {...props} />,
-                }}
-              />
-            </div>
-          )}
-        </div>
-        <div className="d-block d-md-none">
-          <MobileDropdownButton onClick={toggleDropdown}>
-            <i className={`bi ${dropdown ? "bi-x" : "bi-list"}`}></i>
-          </MobileDropdownButton>
-        </div>
-        <div
-          className={`d-md-none ${dropdown ? "d-flex" : "d-none"
-            } w-100 flex-column gap-3 text-white position-absolute start-50 top-100 shadow`}
-          style={{
-            transform: "translateX(-50%)",
-            background: "#0b0c14",
-            padding: "24px 48px",
-            zIndex: 5,
-            borderBottom: "1px solid var(--Stroke-color, rgba(255, 255, 255, 0.20))"
-          }}
-        >
-          {mobileNavbarNavigationItems.map(({ href, label }) => {
-            const isActive = window.location.href === window.location.origin.concat(href)
-
-            return (
-              <MobileLink
-                key={label}
-                href={href}
-                className={isActive && "active"}
-              >
-                {label}
-              </MobileLink>
-            )
-          })}
-          {!props.signedIn && (
-            <button className="sign-in my-3" onClick={props.requestSignIn}>
-              Sign In
-            </button>
-          )}
-          {props.signedIn && (
-            <div>
-              <Widget
-                src="buildhub.near/widget/components.button.join-now"
-                config={{
-                  redirectMap: redirectStore.redirectMap,
-                }}
-                props={{
-                  children: <UserDropdown {...props} />,
-                }}
-              />
-            </div>
-          )}
+        <Navigation />
+        <div className="d-none d-md-flex">
+          <SigninOrJoinNow />
         </div>
       </StyledNavbar>
     </div>
